+++ js/content.js
@@ -1,22 +1,15 @@
-
 var darken = false;
 
-chrome.storage.local.get(['darken_active'], function (result) {
+(async () => {
 
-    darken = result.darken_active;
-    darken = (darken === "true"); //string to boolean
+    await import( chrome.extension.getURL('js/nighteye.js') );
+    chrome.storage.local.get(['darken_active'], function (result) {
 
-    if( darken && !document.getElementById('__nighteye_css') ) {
-      // Activate
-      var body = document.body;
-      var style = document.createElement('link');
-      style.href = chrome.extension.getURL('css/nighteye.css');
-      style.type = 'text/css';
-      style.rel = 'stylesheet';
-      style.id = '__nighteye_css';
+      darken = (result.darken_active === "true"); //string to boolean
+      if( darken && !document.getElementById('__nighteye_css') ) {
+        Nighteye__ext.activate();
+      }
 
-      body.appendChild(style);
-      
-    }
+    });
 
-});
+})();